drop disable/buildbot.patch & disable/mojo.patch
disable/buildbot.patch was needed because we delete the buildbot directory
in the chromium source. Debian doesn't actually use upstream's buildbot stuff,
so at some point mgilbert decided to just delete the whole directory. However, it's
not worth deleting when we also need to maintain a patch to fix the resulting
build issues.

disable/mojo.patch appears to have been for a past build issue when running
with --run-web-tests. If the build works, no need for it any more.