downgrade tsc
The bundled tsc was upgraded from 5.0 to 5.1 in chromium 116, and it
uses language features that aren't available in bullseye's ancient
version of nodejs.

I tried unbunding tsc and pulling in debian's node-typescript package
instead, but chromium's newer code depends in typescript language
features that aren't available in bullseye's ancient version of
node-typscript. So keeping it bundled and downgrading it is the way
to go for now.

4 months of bullseye security support remaining...