refresh/drop patches for bullseye, & grab optional header from libstdc++10
We don't need bullseye/constexpr.patch any more, and downgrade-typescript.patch
needed a refresh.

Also, the optional header that we pulled from libstdc++13 very much does
not work with libstdc++10. So this also changes that, pulling from
bullseye's libstdc++ and applying the is_constructible removals that we need.
(As an added bonus, one function didn't have is_constructible so that could
remain as-is.)