more bookworm build fixes
bookworm/crabbyav1f-macro-scope.patch: fix a weird build error where
the vec!() macro creates a variable that quickly goes out of scope. I'm
not sure why newer rust allows this, as it seems like a legitimate
error, but 🤷.

bookworm/gn-hpp11.patch: newer gn supports .hpp11 extension, bookworms
does not. Just removing the header is a hack, but it's not like we're
shipping it or modifying it or anything; for debian's purposes, not
having an explicit source dependency on the header is fine.

bookworm/rust-split-at-checked.patch: another unstable feature addition
for our older rustc-web.