Fix sporadic partition allocator failure on ppc64el systems
When using a Bookworm userspace and Trixie kernel, the switch to 4k page sizes
in the kernel triggers a fault in the page allocator, whereby too many pages
are allocated for the size of the metadata slot.

While this condition will never be hit on a 64k page kernel, and appears not
to be hit on Trixie/Sid userspace with 4k page kernels, it is still technically
a bug that needs to be fixed.

TODO: Investigate if 64 byte metadata entries are still required on ppc64el
systems